+++ .gitignore
@@ -38,4 +38,3 @@ out/
 
 ### macOS ###
 .DS_STORE
-/gitmessage.txt
