+++ docs/README.md
@@ -35,3 +35,10 @@
 - [x] 사용자는 원하는 시간에 예약할 수 있다. 
 - [x] 사용자 인기 테마 페이지 연결
 - [x] 최근 일주일 내에 예약이 가장 많은 순서대로 테마 10개를 조회할 수 있다. 
+
+# 리팩토링 목록
+
+- [x] 패키지 변경
+- [ ] 인기 테마 조회 로직 변경
+- [ ] 전역 예외 처리 클래스 추가
+- [ ] 테스트 red -> green
