+++ src/main/java/roomescape/controller/MemberController.java
@@ -1,0 +1,13 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MemberController {
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "reservation";
+    }
+}
