+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,58 @@
+package roomescape.controller;
+
+import java.net.URI;
+import java.time.LocalDate;
+import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import roomescape.domain.ReservationTime;
+import roomescape.service.ReservationTimeService;
+
+@RestController
+@RequestMapping("/times")
+public class ReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ReservationTime>> findAll() {
+        return ResponseEntity.ok(reservationTimeService.findAll());
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
+        ReservationTime createdTime = reservationTimeService.save(reservationTime);
+        return ResponseEntity.created(URI.create("/times/" + createdTime.id())).body(createdTime);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable long id) {
+        reservationTimeService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/user")
+    public ResponseEntity<List<ReservationTime>> availableTime(@RequestParam String date, @RequestParam long themeId){
+        LocalDate parse = LocalDate.parse(date);
+        return ResponseEntity.ok(reservationTimeService.available(parse,themeId));
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<Void> handleIllegalArgumentException(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().build();
+    }
+}
