+++ src/main/java/roomescape/controller/request/ReservationRequest.java
@@ -1,0 +1,6 @@
+package roomescape.controller.request;
+
+import java.time.LocalDate;
+
+public record ReservationRequest(String name, LocalDate date, long timeId, long themeId) {
+}
