+++ src/main/java/roomescape/domain/RankTheme.java
@@ -1,0 +1,4 @@
+package roomescape.domain;
+
+public record RankTheme(String name,String thumbnail,String description) {
+}
