+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,0 +1,34 @@
+package roomescape.domain;
+
+import java.time.LocalDate;
+
+public record Reservation(long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+
+    public Reservation(long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        this.id = id;
+        this.name = validateName(name);
+        //LocalDateTime dateTime = validateDateTime(date, time.startAt());
+        //this.date = dateTime.toLocalDate();
+        this.date=date;
+        this.time = time;
+        this.theme = theme;
+    }
+
+    private String validateName(String name) {
+        if (name.isBlank()) {
+            throw new IllegalArgumentException("Name cannot be empty");
+        }
+        if (name.length() > 10) {
+            throw new IllegalArgumentException("Name cannot exceed 10 characters");
+        }
+        return name;
+    }
+
+//    private LocalDateTime validateDateTime(LocalDate date, LocalTime time) {
+//        LocalDateTime dateTime = LocalDateTime.of(date, time);
+//        if (dateTime.isBefore(LocalDateTime.now())) {
+//            throw new IllegalArgumentException("Cannot create a reservation for a past date and time.");
+//        }
+//        return dateTime;
+//    }
+}
