+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,0 +1,13 @@
+package roomescape.domain;
+
+import java.time.LocalTime;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+public record ReservationTime(long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
+    public ReservationTime {
+        if (startAt == null) {
+            throw new IllegalArgumentException("Start time cannot be null");
+        }
+    }
+}
