+++ src/main/java/roomescape/domain/Theme.java
@@ -1,0 +1,4 @@
+package roomescape.domain;
+
+public record Theme(long id, String name, String description, String thumbnail) {
+}
