--- src/main/java/roomescape/member/MemberController.java
@@ -1,13 +1,0 @@
-package roomescape.member;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MemberController {
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "reservation";
-    }
-}
