--- src/main/java/roomescape/ranking/RankTheme.java
@@ -1,4 +1,0 @@
-package roomescape.ranking;
-
-public record RankTheme(String name,String thumbnail,String description) {
-}
