--- src/main/java/roomescape/reservation/request/ReservationRequest.java
@@ -1,6 +1,0 @@
-package roomescape.reservation.request;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(String name, LocalDate date, long timeId, long themeId) {
-}
