--- src/main/java/roomescape/theme/domain/Theme.java
@@ -1,4 +1,0 @@
-package roomescape.theme.domain;
-
-public record Theme(long id, String name, String description, String thumbnail) {
-}
