--- src/main/java/roomescape/time/domain/ReservationTime.java
@@ -1,13 +1,0 @@
-package roomescape.time.domain;
-
-import java.time.LocalTime;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-public record ReservationTime(long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
-    public ReservationTime {
-        if (startAt == null) {
-            throw new IllegalArgumentException("Start time cannot be null");
-        }
-    }
-}
