--- src/main/java/roomescape/user/UserController.java
@@ -1,17 +1,0 @@
-package roomescape.user;
-
-import org.springframework.web.bind.annotation.RestController;
-
-import roomescape.reservation.service.ReservationService;
-
-@RestController
-public class UserController {
-
-    private final ReservationService reservationService;
-
-    public UserController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-
-}
